+++ utils/src/lib.rs
@@ -5,23 +5,23 @@ use std::env;
 use std::fs;
 use std::io::Error;
 
-pub fn read_integers_from_param() -> Vector<i32> {
+pub fn read_integers_from_param() -> Vector<i64> {
     read_integers(&input_file_from_param()).expect("Failed to read integers fro input file")
 }
 
-pub fn read_integers(path: &str) -> Result<Vector<i32>, Error> {
+pub fn read_integers(path: &str) -> Result<Vector<i64>, Error> {
     read_strings(path).map(|lines| {
         lines
             .iter()
-            .map(|line| line.parse::<i32>().unwrap())
+            .map(|line| line.parse::<i64>().unwrap())
             .collect()
     })
 }
 
-pub fn parse_integers(input: &str) -> Vector<i32> {
+pub fn parse_integers(input: &str) -> Vector<i64> {
     input
         .lines()
-        .map(|line| line.parse::<i32>().unwrap())
+        .map(|line| line.parse::<i64>().unwrap())
         .collect()
 }
